Update README.md
Significant updates made including:
- Updated main index page 
- Discord authentication has been implemented and specific registration removed
- Filtering added to cards list view
- Decks view export features added 
- Added ability for users to delete current decks
- Deck builder token gate has been removed, all public users can access the deck builder
- Checklist created for authenticated users

In addition to the current updates, we've also been working with card sales sites to implement the ability to reference cards and pull live data such as pricing and trends.  Once the data is publicly available we will update the cards and decks pages for easy purchasing of live cards.